+++ src/components/common/Post/Post.tsx
@@ -1,11 +1,10 @@
 import { MouseEvent } from "react";
 
-import { useMutation, useQueryClient } from "@tanstack/react-query";
-import { AxiosError } from "axios";
 import classNames from "classnames/bind";
 
 import Link from "next/link";
 
+import { useLikeMutation } from "@/apis/post";
 import styles from "@/components/common/Post/Post.module.scss";
 import { ROUTE } from "@/constants/route";
 import Calendar from "@/icons/calendar.svg";
@@ -16,73 +15,26 @@ import PinkHeart from "@/icons/pink_heart.svg";
 import { PostType } from "@/types/post";
 import { formatDateString } from "@/utils";
 
-import postLikes from "./apis/postLikes";
 import PostLabel from "./PostLabel/PostLabel";
-import openToast from "../Toast/features/openToast";
 
 const cn = classNames.bind(styles);
 
 interface PostProps {
   data: PostType;
+  children?: React.ReactNode;
 }
 
-interface ErrorResponse {
-  error: {
-    message: string;
-  };
-}
-
-export default function Post({ data }: PostProps) {
-  const queryClient = useQueryClient();
-
+export function Post({ data, children }: PostProps) {
   const {
     assistance: { assistanceEndTime, assistanceStartTime, assistanceType },
     disabilityType,
     district,
     id,
-    isLiked,
     postStatus,
     postType,
     schedule: { endDate, scheduleType, startDate },
     title,
   } = data;
-
-  const { mutate } = useMutation({
-    mutationFn: () => postLikes(id),
-    onMutate: async () => {
-      if (localStorage.getItem("accessToken")) {
-        await queryClient.cancelQueries({ queryKey: ["takerPost"] });
-
-        const previousTodos = queryClient.getQueryData(["takerPost"]);
-
-        queryClient.setQueryData(["takerPost"], (postList: PostType[]) =>
-          postList.map((item) => {
-            if (item.id === id) {
-              return { ...item, isLiked: !item.isLiked };
-            }
-            return item;
-          }),
-        );
-        return { previousTodos };
-      }
-    },
-    onError: (error: AxiosError<ErrorResponse>, newTodo, context) => {
-      queryClient.setQueryData(["takerPost"], context?.previousTodos);
-      if (error.response?.status === 401) {
-        openToast("warn", "로그인이 필요한 서비스입니다.");
-      } else {
-        openToast("warn", "에러가 발생하였습니다.");
-      }
-    },
-    onSettled: () => {
-      queryClient.invalidateQueries({ queryKey: ["takerPost"] });
-    },
-  });
-
-  const handleHeartClick = (event: MouseEvent<SVGSVGElement>) => {
-    event.preventDefault();
-    mutate();
-  };
 
   return (
     <Link
@@ -101,11 +53,7 @@ export default function Post({ data }: PostProps) {
         >
           {postStatus === "RECRUITING" ? "모집중" : "모집완료"}
         </p>
-        {isLiked ? (
-          <PinkHeart onClick={handleHeartClick} width={32} height={32} className={cn("heart")} />
-        ) : (
-          <Heart onClick={handleHeartClick} width={32} height={32} className={cn("heart")} />
-        )}
+        {children}
         <div className={cn("contentBox")}>
           <div className={cn("box")}>
             <p className={cn("title")}>{title}</p>
@@ -134,5 +82,30 @@ export default function Post({ data }: PostProps) {
         {disabilityType !== "없음" && <PostLabel text={disabilityType} />}
       </div>
     </Link>
+  );
+}
+
+interface PostHeartProps {
+  id: number;
+  isLiked: boolean;
+  queryKey: string[];
+}
+
+export function PostHeart({ id, isLiked, queryKey }: PostHeartProps) {
+  const { mutate } = useLikeMutation({ id, queryKey: queryKey });
+
+  const handleHeartClick = (event: MouseEvent<SVGSVGElement>) => {
+    event.preventDefault();
+    mutate();
+  };
+
+  return (
+    <>
+      {isLiked ? (
+        <PinkHeart onClick={handleHeartClick} width={32} height={32} className={cn("heart")} />
+      ) : (
+        <Heart onClick={handleHeartClick} width={32} height={32} className={cn("heart")} />
+      )}
+    </>
   );
 }
