+++ src/types/error/index.ts
@@ -1,0 +1,5 @@
+export interface ErrorResponse {
+  error: {
+    message: string;
+  };
+}
